// ClientApp/src/store/toolbox/toolbox.types.ts
export enum TOOLBOX_ACTION_TYPES {
    SET_IS_PANEL_OPEN = 'toolbox/SET_IS_PANEL_OPEN',
    SET_IS_NOTE_OPEN = 'toolbox/SET_IS_NOTE_OPEN',
    SET_IS_ARTOO_OPEN = 'toolbox/SET_IS_ARTOO_OPEN',
    SET_IS_CHATBOT_OPEN = 'toolbox/SET_IS_CHATBOT_OPEN',
    SET_IS_SEARCHBAR_OPEN = 'toolbox/SET_IS_SEARCHBAR_OPEN',
    SET_IS_EDITOR_OPEN = 'toolbox/SET_IS_EDITOR_OPEN',
    SET_IS_TASKS_OPEN = 'toolbox/SET_IS_TASKS_OPEN',
    SET_IS_RESET_OPEN = 'toolbox/SET_IS_RESET_OPEN',
    SET_IS_MESSAGES_OPEN = 'toolbox/SET_IS_MESSAGES_OPEN',
    SET_IS_COMMUNITIES_OPEN = 'toolbox/SET_IS_COMMUNITIES_OPEN'
}